GameState.clone() is VERY slow because it copies all the location objects in the board.  This is fine for normal game play but not ok when you need to create lots and lots of Nodes that contains a state when doing search-based AI.
So, I changed GameState so that it contains a third inventory for a neutral player that contains all the grass and food.  I had to update some methods in GameState.py and Game.py to support this.

With this change all the data on the board is now also in the inventories.  So I was able to add a fastclone() method to GameState that copies the inventories but ignores the board.

Then I had to change the AIPlayerUtils.py to NOT use the board.  Ever.  That way they are still quite effective for search-based AIs.  

I feel like it's pretty stable but it's a big change and I probably missed a few bugs. :/