Added new rule:  queen stays in her territory
To prevent easy zerg rush tactics, the game now will not allow the queen
to leave her own territory.  I modified Game.py to enforce this and
then modified AIPlayerUtils.py to reflect this.  The default (Random)
AI (AIPlayer.py) was modified to use AIPlayerUtils.py to select its
random moves since this was a) simpler and b) was the easiest way to
prevent it from moving the queen outside of her territory.